+++ domain/src/main/java/com/mnassa/domain/interactor/impl/PostsInteractorImpl.kt
@@ -188,7 +188,7 @@ class PostsInteractorImpl(private val postsRepository: PostsRepository,
         postsRepository.updateUserRecommendation(postId, accountId, text)
     }
 
-    override suspend fun getDefaultExpirationDays(): Int = postsRepository.getDefaultExpirationDays()
+    override suspend fun getDefaultExpirationDays(): Long = postsRepository.getDefaultExpirationDays()
 
     private companion object {
         private const val SEND_VIEWED_ITEMS_BUFFER_DELAY = 1_000L
